--- RecoTracker/CkfPattern/interface/MaxHitsTrajectoryFilter.h
@@ -1,2 +1,0 @@
-#warning header has moved to TrackingTools/TrajectoryFiltering/interface/MaxHitsTrajectoryFilter.h
-#include "TrackingTools/TrajectoryFiltering/interface/MaxHitsTrajectoryFilter.h"
