--- RecoTracker/CkfPattern/interface/MinPtTrajectoryFilter.h
@@ -1,2 +1,0 @@
-#warning header has moved to TrackingTools/TrajectoryFiltering/interface/MinPtTrajectoryFilter.h
-#include "TrackingTools/TrajectoryFiltering/interface/MinPtTrajectoryFilter.h"
