--- RecoTracker/CkfPattern/interface/RegionalTrajectoryFilter.h
@@ -1,2 +1,0 @@
-#warning header has moved to TrackingTools/TrajectoryFiltering/interface/RegionalTrajectoryFilter.h
-#include "TrackingTools/TrajectoryFiltering/interface/RegionalTrajectoryFilter.h"
