--- RecoTracker/CkfPattern/interface/TempTrajectory.h
@@ -1,2 +1,0 @@
-#warning header has moved to TrackingTools/PatternTools/interface/TempTrajectory.h
-#include "TrackingTools/PatternTools/interface/TempTrajectory.h"
