--- RecoTracker/CkfPattern/interface/TrajectoryFilter.h
@@ -1,2 +1,0 @@
-#warning header has moved to TrackingTools/TrajectoryFiltering/interface/TrajectoryFilter.h
-#include "TrackingTools/TrajectoryFiltering/interface/TrajectoryFilter.h"
