--- RecoTracker/CkfPattern/plugins/CkfTrackCandidateMaker.cc
@@ -1,1 +1,0 @@
-#include "RecoTracker/CkfPattern/interface/CkfTrackCandidateMaker.h"
